+++ packages/runtime/plugin-runtime/src/ssr/serverRender/renderToString/buildHtml.ts
@@ -1,3 +1,11 @@
+import { unsafeReplace } from '../utils';
+import {
+  HTML_PLACEHOLDER,
+  SSR_DATA_PLACEHOLDER,
+  CHUNK_JS_PLACEHOLDER,
+  CHUNK_CSS_PLACEHOLDER,
+} from '../constants';
+
 export type BuildHtmlCb = (tempalte: string) => string;
 
 export function buildHtml(template: string, callbacks: BuildHtmlCb[]) {
@@ -5,21 +13,20 @@ export function buildHtml(template: string, callbacks: BuildHtmlCb[]) {
 }
 
 export function createReplaceHtml(html: string): BuildHtmlCb {
-  const HTML_REG = /<!--<\?-\s*html\s*\?>-->/;
-  return (template: string) => template.replace(HTML_REG, html);
+  return (template: string) => unsafeReplace(template, HTML_PLACEHOLDER, html);
 }
 
 export function createReplaceSSRDataScript(data: string): BuildHtmlCb {
-  const SSR_DATA_REG = /<!--<\?-\s*SSRDataScript\s*\?>-->/;
-  return (template: string) => template.replace(SSR_DATA_REG, data);
+  return (template: string) =>
+    unsafeReplace(template, SSR_DATA_PLACEHOLDER, data);
 }
 
 export function createReplaceChunkJs(js: string): BuildHtmlCb {
-  const CHUNK_JS_REG = /<!--<\?-\s*chunksMap\.js\s*\?>-->/;
-  return (template: string) => template.replace(CHUNK_JS_REG, js);
+  return (template: string) =>
+    unsafeReplace(template, CHUNK_JS_PLACEHOLDER, js);
 }
 
 export function createReplaceChunkCss(css: string): BuildHtmlCb {
-  const CHUNK_CSS_REG = /<!--<\?-\s*chunksMap\.css\s*\?>-->/;
-  return (template: string) => template.replace(CHUNK_CSS_REG, css);
+  return (template: string) =>
+    unsafeReplace(template, CHUNK_CSS_PLACEHOLDER, css);
 }