+++ packages/server/server/src/index.ts
@@ -1,9 +1,13 @@
-import { DevServer as Server } from './server';
+import {
+  DevServer as Server,
+  DevServerForRsbuild as ServerForRsbuild,
+} from './server';
 import type { ModernDevServerOptions } from './types';
 
-export { Server };
+export { Server, ServerForRsbuild };
 export type { ModernDevServerOptions };
 
+// TODO: it seems not used in any pkgs?
 export default (options: ModernDevServerOptions): Promise<Server> => {
   if (options == null) {
     throw new Error('can not start server without options');