+++ packages/cli/plugin-lint/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @modern-js/plugin-lint
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/utils@2.42.2
+- @modern-js/tsconfig@2.42.2
+
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.42.0
+- @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes