+++ packages/cli/plugin-proxy/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @modern-js/plugin-proxy
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/utils@2.42.2
+
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes