+++ packages/devtools/client/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @modern-js/devtools-client
 
+## 2.42.2
+
+### Patch Changes
+
+- 07c56c0: chore: bump rsbuild to 0.2.3
+
+  chore: 升级 rsbuild 到 0.2.3
+
+- 859fca0: feat(devtools): set z-index to avoid be covered
+  feat(devtools): 设置 z-index 以避免被覆盖
+
+## 2.42.1
+
+## 2.42.0
+
 ## 2.41.0
 
 ## 2.40.0