+++ packages/document/module-doc/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @modern-js/module-tools-docs
 
+## 2.42.2
+
+### Patch Changes
+
+- 24e25e2: docs(module-tools): update example link
+  docs(module-tools): 更新示例链接
+
+## 2.42.1
+
+### Patch Changes
+
+- 7d327e5: docs: change builderCofnig to builderConfig
+
+## 2.42.0
+
 ## 2.41.0
 
 ## 2.40.0