+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @modern-js/generator-common
 
+## 3.3.3
+
+### Patch Changes
+
+- ad2862b: fix tips typo of testingPlugin, "testPlugin" -> "testingPlugin"
+- fa731a7: chore: bump codesmith to 2.3.2
+  chore: 升级 codesmith 版本到 2.3.2
+  - @modern-js/plugin-i18n@2.42.2
+
+## 3.3.2
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.42.1
+
+## 3.3.1
+
+### Patch Changes
+
+- b182eb2: chore: bump codesmith v2.3.1 to show timing
+  chore: 升级 codesmith 到 v2.3.1, 用于展示执行时间
+  - @modern-js/plugin-i18n@2.42.0
+
 ## 3.3.0
 
 ### Minor Changes