+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @modern-js/generator-utils
 
+## 3.3.3
+
+### Patch Changes
+
+- fa731a7: chore: bump codesmith to 2.3.2
+  chore: 升级 codesmith 版本到 2.3.2
+- Updated dependencies [ad2862b]
+- Updated dependencies [fa731a7]
+  - @modern-js/generator-common@3.3.3
+  - @modern-js/utils@2.42.2
+  - @modern-js/plugin-i18n@2.42.2
+
+## 3.3.2
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.42.1
+- @modern-js/utils@2.42.1
+- @modern-js/generator-common@3.3.2
+
+## 3.3.1
+
+### Patch Changes
+
+- b182eb2: chore: bump codesmith v2.3.1 to show timing
+  chore: 升级 codesmith 到 v2.3.1, 用于展示执行时间
+- Updated dependencies [b182eb2]
+  - @modern-js/generator-common@3.3.1
+  - @modern-js/plugin-i18n@2.42.0
+  - @modern-js/utils@2.42.0
+
 ## 3.3.0
 
 ### Patch Changes