+++ packages/generator/generators/module-test-generator/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @modern-js/module-test-generator
 
+## 3.3.3
+
+### Patch Changes
+
+- fa731a7: chore: bump codesmith to 2.3.2
+  chore: 升级 codesmith 版本到 2.3.2
+  - @modern-js/utils@2.42.2
+
+## 3.3.2
+
+### Patch Changes
+
+- @modern-js/utils@2.42.1
+
+## 3.3.1
+
+### Patch Changes
+
+- @modern-js/utils@2.42.0
+
 ## 3.3.0
 
 ### Patch Changes