+++ packages/module/plugin-module-banner/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @modern-js/plugin-module-banner
 
+## 2.42.2
+
+### Patch Changes
+
+- Updated dependencies [ffeb83b]
+- Updated dependencies [4bce612]
+- Updated dependencies [923fee7]
+- Updated dependencies [eda89cb]
+- Updated dependencies [e4dd353]
+- Updated dependencies [82e0af0]
+  - @modern-js/module-tools@2.42.2
+
+## 2.42.1
+
+### Patch Changes
+
+- Updated dependencies [fef529c]
+  - @modern-js/module-tools@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- Updated dependencies [c2fa7c4]
+- Updated dependencies [5d2f474]
+- Updated dependencies [6200cf2]
+  - @modern-js/module-tools@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes