+++ packages/server/create-request/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @modern-js/create-request
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.42.2
+- @modern-js/utils@2.42.2
+
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- Updated dependencies [103cf92]
+  - @modern-js/runtime-utils@2.42.0
+  - @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes