+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @modern-js/server-utils
 
+## 2.42.2
+
+### Patch Changes
+
+- 07c56c0: chore: bump rsbuild to 0.2.3
+
+  chore: 升级 rsbuild 到 0.2.3
+
+  - @modern-js/utils@2.42.2
+  - @modern-js/babel-plugin-module-resolver@2.42.2
+  - @modern-js/babel-compiler@2.42.2
+
+## 2.42.1
+
+### Patch Changes
+
+- @modern-js/babel-plugin-module-resolver@2.42.1
+- @modern-js/babel-compiler@2.42.1
+- @modern-js/utils@2.42.1
+
+## 2.42.0
+
+### Patch Changes
+
+- @modern-js/babel-plugin-module-resolver@2.42.0
+- @modern-js/babel-compiler@2.42.0
+- @modern-js/utils@2.42.0
+
 ## 2.41.0
 
 ### Patch Changes